corrected long-time world animation issue Updated UI

+++ common/src/main/java/net/dialingspoon/speedcap/interfaces/LivingEntityInterface.java
@@ -3,7 +3,7 @@ package net.dialingspoon.speedcap.interfaces;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 
 public interface LivingEntityInterface {
-    float speedcap$getSailTick();
+    long speedcap$getSailTick();
 
     boolean speedcap$sailDirection();
